pull

+++ Realm/RLMConstants.h
@@ -23,7 +23,7 @@
 
  See [Realm Models](http://realm.io/docs/cocoa/latest/#models)
  */
-// Make sure numbers match those in <tightdb/data_type.hpp>
+// Make sure numbers match those in <realm/data_type.hpp>
 typedef NS_ENUM(int32_t, RLMPropertyType) {
     ////////////////////////////////
     // Primitive types
@@ -111,8 +111,8 @@ typedef NS_ENUM(NSInteger, RLMError) {
     RLMErrorIncompatibleLockFile  = 8,
 };
 
-// Schema version used for unitialized Realms
-extern const NSUInteger RLMNotVersioned;
+// Schema version used for uninitialized Realms
+extern const uint64_t RLMNotVersioned;
 
 extern NSString * const RLMErrorDomain;
